refactor: extract deterministic tiers and helpers from qa/workflow.ts
